requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
numpy>=1.24
faiss-cpu>=1.7.4
mistralai>=1.0.0
